Concurrent upload check

+++ src/main/java/at/yeoman/photobackup/server/service/RequestHandler.java
@@ -18,6 +18,8 @@ import java.nio.channels.FileLock;
 import java.security.MessageDigest;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
 @RestController
@@ -28,6 +30,8 @@ public class RequestHandler {
     private final File assetDirectory;
     private final File uploadDirectory;
     private final File photoDirectory;
+
+    private final Set<Checksum> checksumsUploading = ConcurrentHashMap.newKeySet();
 
     @Autowired
     RequestHandler(Core core) throws IOException {
@@ -105,41 +109,50 @@ public class RequestHandler {
                 return error("Illegal checksum string from path [" + checksumString + "]",
                         HttpStatus.BAD_REQUEST);
             }
-            String fileName = fileNameForChecksum(checksumFromPath);
-            File uploadTarget = new File(uploadDirectory, fileName);
-            final Checksum calculatedChecksum = writeFile(bodyStream, uploadTarget);
-            if (!calculatedChecksum.equals(checksumFromPath)) {
-                return error("Checksum mismatch - checksum from request path: " + checksumFromPath +
-                        ", calculated: " + calculatedChecksum,
+            if (!checksumsUploading.add(checksumFromPath)) {
+                return error("Attempt to concurrently upload file of same checksum " + checksumFromPath,
                         HttpStatus.EXPECTATION_FAILED);
             }
-            File renamedTarget = new File(photoDirectory, fileName);
-            if (renamedTarget.exists()) {
-                if (!renamedTarget.isFile()) {
-                    return error("renamed target [" + renamedTarget.getCanonicalPath() + "] is not a file",
+            try {
+                String fileName = fileNameForChecksum(checksumFromPath);
+                File uploadTarget = new File(uploadDirectory, fileName);
+                final Checksum calculatedChecksum = writeFile(bodyStream, uploadTarget);
+                if (!calculatedChecksum.equals(checksumFromPath)) {
+                    return error("Checksum mismatch - checksum from request path: " + checksumFromPath +
+                                    ", calculated: " + calculatedChecksum,
                             HttpStatus.EXPECTATION_FAILED);
-                } else if (renamedTarget.length() > uploadTarget.length()) {
-                    return error("renamed target [" + renamedTarget.getCanonicalPath() + "] length [" + renamedTarget.length() + "]" +
-                            " > upload target [" + uploadTarget.getCanonicalPath() + "] length [" + uploadTarget.length() + "]",
-                            HttpStatus.EXPECTATION_FAILED);
-                } else if (renamedTarget.length() == uploadTarget.length() && checksumForFile(uploadTarget).equals(checksumFromPath)) {
-                    if (!uploadTarget.delete()) {
-                        log.error("Unable to delete upload target while retaining existing backup [" + uploadTarget.getCanonicalPath() + "]");
-                    }
-                    return success("Retaining matching file [" + renamedTarget.getCanonicalPath());
-                } else {
-                    if (!renamedTarget.delete()) {
-                        return error("Unable to delete non-matching renamed target [" + renamedTarget.getCanonicalPath() + "]",
-                                HttpStatus.INTERNAL_SERVER_ERROR);
-                    }
-                    if (!uploadTarget.renameTo(renamedTarget)) {
-                        return error("Unable to rename [" + uploadTarget.getCanonicalPath() + "] to [" + renamedTarget.getCanonicalPath() + "]",
-                                HttpStatus.INTERNAL_SERVER_ERROR);
+                }
+                File renamedTarget = new File(photoDirectory, fileName);
+                if (renamedTarget.exists()) {
+                    if (!renamedTarget.isFile()) {
+                        return error("renamed target [" + renamedTarget.getCanonicalPath() + "] is not a file",
+                                HttpStatus.EXPECTATION_FAILED);
+                    } else if (renamedTarget.length() > uploadTarget.length()) {
+                        return error("renamed target [" + renamedTarget.getCanonicalPath() + "] length [" + renamedTarget.length() + "]" +
+                                        " > upload target [" + uploadTarget.getCanonicalPath() + "] length [" + uploadTarget.length() + "]",
+                                HttpStatus.EXPECTATION_FAILED);
+                    } else if (renamedTarget.length() == uploadTarget.length() && checksumForFile(uploadTarget).equals(checksumFromPath)) {
+                        if (!uploadTarget.delete()) {
+                            log.error("Unable to delete upload target while retaining existing backup [" + uploadTarget.getCanonicalPath() + "]");
+                        }
+                        return success("Retaining matching file [" + renamedTarget.getCanonicalPath());
+                    } else {
+                        if (!renamedTarget.delete()) {
+                            return error("Unable to delete non-matching renamed target [" + renamedTarget.getCanonicalPath() + "]",
+                                    HttpStatus.INTERNAL_SERVER_ERROR);
+                        }
+                        if (!uploadTarget.renameTo(renamedTarget)) {
+                            return error("Unable to rename [" + uploadTarget.getCanonicalPath() + "] to [" + renamedTarget.getCanonicalPath() + "]",
+                                    HttpStatus.INTERNAL_SERVER_ERROR);
+                        }
                     }
                 }
+                return error(calculatedChecksum.toString(), HttpStatus.OK);
+            } finally {
+                checksumsUploading.remove(checksumFromPath);
             }
-            return error(calculatedChecksum.toString(), HttpStatus.OK);
         } catch (Exception exception) {
+            bodyStream.close();
             log.error(exception.getMessage(), exception);
             return error("Error message: [" + exception.getMessage() + "]",
                     HttpStatus.EXPECTATION_FAILED);
